dimension marshal tests wrapping up
